Update to use new angular module methodology
Using njInject annotation now instead of previous manual method combined with grunt.
This module is meant to be used with webpack so there is no need for grunt.
Stripped out the tests; because this is a special use case and I don't have time to make them function.

Instead of building in Mousetrap, converted Mousetrap into a module and loading it that way.
Moved the "monkeypatch" of Mousetrap into the newly angular modular converted Mousetrap itself.

TODO: Fix pause and unpause as I believe I've broken them in porting them into Mousetrap.